Split ChunkFileLoader.* to Block*, Item* and Player* loader files
