+++ Bloki.vcxproj.filters
@@ -103,6 +103,12 @@
     <Filter Include="Pliki źródłowe\engine\qxml">
       <UniqueIdentifier>{c6273c8e-fab7-4730-8a03-58ef962a70da}</UniqueIdentifier>
     </Filter>
+    <Filter Include="Pliki źródłowe\game\save loading">
+      <UniqueIdentifier>{3d8e0069-4fa3-4004-b766-9048aab47177}</UniqueIdentifier>
+    </Filter>
+    <Filter Include="Pliki nagłówkowe\game\save loading">
+      <UniqueIdentifier>{cac7c13a-18f9-4685-8319-c419eda9831a}</UniqueIdentifier>
+    </Filter>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="src\Bloki.cpp">
@@ -148,9 +154,6 @@
       <Filter>Pliki źródłowe\game\entity</Filter>
     </ClCompile>
     <ClCompile Include="src\game\world\LoadedChunks.cpp">
-      <Filter>Pliki źródłowe\game\world</Filter>
-    </ClCompile>
-    <ClCompile Include="src\game\world\ChunkFileLoader.cpp">
       <Filter>Pliki źródłowe\game\world</Filter>
     </ClCompile>
     <ClCompile Include="src\game\core\Player.cpp">
@@ -312,6 +315,18 @@
     <ClCompile Include="src\game\entity\DroppedItemRepository.cpp">
       <Filter>Pliki źródłowe\game\entity</Filter>
     </ClCompile>
+    <ClCompile Include="src\game\save_loading\WholeSaveLoader.cpp">
+      <Filter>Pliki źródłowe\game\save loading</Filter>
+    </ClCompile>
+    <ClCompile Include="src\game\save_loading\BlockDataLoader.cpp">
+      <Filter>Pliki źródłowe\game\save loading</Filter>
+    </ClCompile>
+    <ClCompile Include="src\game\save_loading\ItemDataLoader.cpp">
+      <Filter>Pliki źródłowe\game\save loading</Filter>
+    </ClCompile>
+    <ClCompile Include="src\game\save_loading\PlayerDataLoader.cpp">
+      <Filter>Pliki źródłowe\game\save loading</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="src\engine\util\Debug.hpp">
@@ -364,9 +379,6 @@
     </ClInclude>
     <ClInclude Include="src\game\entity\Entity.hpp">
       <Filter>Pliki nagłówkowe\game\entity</Filter>
-    </ClInclude>
-    <ClInclude Include="src\game\world\ChunkFileLoader.hpp">
-      <Filter>Pliki nagłówkowe\game\world</Filter>
     </ClInclude>
     <ClInclude Include="src\game\core\Player.hpp">
       <Filter>Pliki nagłówkowe\game\core</Filter>
@@ -535,6 +547,18 @@
     </ClInclude>
     <ClInclude Include="src\game\entity\DroppedItemRepository.hpp">
       <Filter>Pliki nagłówkowe\game\entity</Filter>
+    </ClInclude>
+    <ClInclude Include="src\game\save_loading\WholeSaveLoader.hpp">
+      <Filter>Pliki nagłówkowe\game\save loading</Filter>
+    </ClInclude>
+    <ClInclude Include="src\game\save_loading\BlockDataLoader.hpp">
+      <Filter>Pliki nagłówkowe\game\save loading</Filter>
+    </ClInclude>
+    <ClInclude Include="src\game\save_loading\ItemDataLoader.hpp">
+      <Filter>Pliki nagłówkowe\game\save loading</Filter>
+    </ClInclude>
+    <ClInclude Include="src\game\save_loading\PlayerDataLoader.hpp">
+      <Filter>Pliki nagłówkowe\game\save loading</Filter>
     </ClInclude>
   </ItemGroup>
 </Project>
